Remove `family.family_clade` association from dbmodels.
`family.family_clade` was defined as a hasOne, duplicating the more
correct `hasMany` association `family.clades` that points to
`dfamTaxdbModel`. The fact that this retrieved the expected results
appears to be an accident.

Using the `clades` association slightly simplifies the appearance of the
code for filtering for families, but requires working around a sequelize
bug.